Edit picture, clean up code in MegaBot.java
