+++ data/megabot.txt
@@ -1,4 +1,6 @@
 T | 0 | iP
-T | 0 | A-AI
+T | 1 | A-AI
 T | 0 | A-UserGuide
 D | 0 | CS2100 iP submission | 2025-09-19
+T | 1 | A-BetterGui
+T | 1 | A-MoreTesting
